refactor(planning): sortir les FormRequests Planning du module HR — isolation #6588 lot 1 (Closes #7154)
Les 3 FormRequests des endpoints Planning (store/update schedule, assign employees)
résidaient dans Modules/HR/Interfaces/Api/V1/Requests et n'étaient consommées que par
Planning (ScheduleController). Déplacement vers Modules/Planning/.../Requests,
namespace mis à jour, allowlist d'isolation décrémentée (Modules/Planning -> Modules/HR
supprimée). Aucun changement de comportement (mêmes règles de validation, mêmes routes).
