+++ api/app/Modules/Planning/Interfaces/Api/V1/Controllers/ScheduleController.php
@@ -8,9 +8,9 @@ use App\Core\Auth\Domain\Models\Employee;
 use App\Core\Tenant\Infrastructure\Services\TenantCacheService;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\Api\V1\ScheduleResource;
-use App\Modules\HR\Interfaces\Api\V1\Requests\AssignScheduleEmployeesRequest;
-use App\Modules\HR\Interfaces\Api\V1\Requests\StoreScheduleRequest;
-use App\Modules\HR\Interfaces\Api\V1\Requests\UpdateScheduleRequest;
+use App\Modules\Planning\Interfaces\Api\V1\Requests\AssignScheduleEmployeesRequest;
+use App\Modules\Planning\Interfaces\Api\V1\Requests\StoreScheduleRequest;
+use App\Modules\Planning\Interfaces\Api\V1\Requests\UpdateScheduleRequest;
 use App\Modules\Planning\Application\Actions\CreateSchedule;
 use App\Modules\Planning\Application\Actions\DeleteSchedule;
 use App\Modules\Planning\Application\Actions\UpdateSchedule;
