+++ api/app/Modules/Planning/Interfaces/Api/V1/Requests/AssignScheduleEmployeesRequest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Modules\HR\Interfaces\Api\V1\Requests;
+namespace App\Modules\Planning\Interfaces\Api\V1\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
